Updated the maxtime for running in windows
In windows , if the year parameter of the datetime.datetime class is greater than 3001 , it returns OSError: [Errno 22] Invalid argument for Windows 10 operating system. It works fine in Linux. It could be a bug in the python. Setting the year value for max_time to 3001 will fix this error for windows.